Update stellar statistics
