+++ include/stellar/options/eps_match_options.hpp
@@ -9,7 +9,7 @@ namespace stellar
 struct EPSMatchOptions
 {
     stellar::utils::fraction epsilon{5, 100}; // maximal error rate
-    int minLength{100}; // minimal length of an epsilon-match
+    unsigned minLength{100}; // minimal length of an epsilon-match
 };
 
 } // namespace stellar
