+++ test/data/out.gff
@@ -1,0 +1,3 @@
+1	Stellar	eps-matches	628	813	95.2127	+	.	1;seq2Range=446,632;cigar=2M1I172M1I1M1D10M;mutations=3A,46C,63G,96C,105T,135G,168C,176C
+1	Stellar	eps-matches	580	707	95.3125	+	.	1;seq2Range=1056,1183;cigar=128M;mutations=12A,22A,33T,39T,52A,63G
+1	Stellar	eps-matches	469	582	95.614	+	.	1;seq2Range=60,173;cigar=114M;mutations=15G,58C,60A,87A,90G
