+++ test/data/query_e0.05.fasta
@@ -1,0 +1,2 @@
+>1
+TCAAAAACCCCGTGTCCCGACAAACTCCGATGTAAGAGCGACTTAAGCGATGACCGGTCAGATAATAGAAGGCGTGAATCCGGCGTTACACCCATAGGGGAATAACGCCGAAATTGCTAGTTCTCGATAATTGCCAGTAATGCATAACGGCATCTGCCTGGGTGGGGAATTGGCCGGCCAGTATAGCGGATCGGTTAAATAATCCAAGGACGTTTCCGCTTAGATTTTGAGGACTTTAACAGGTTTTCGTCTGGCAGAATCATGTCCTTACTGCTTATCCAGGTCTTTTAAAGTTCGTGTCAGACTTTGGTCCCAAGCGACTCCAAACGGAGAACCGTACGAATATCTACCCAGTCCTACGCAAAAATAACTAAGTGGATGACAAGGGTACAACGTTATATAGAGCTTATAAGTGTCACCTAAGCGGCTTCTTAGTGGGACTTCGATAACCCATCTGGTCATCGGGGATCCGATGGCATCGCCAGGTATTCCGCCCCTCCATGAGAAGAAAACAGCTCGGATAACGGTCAAACCGGCAGACGGTTAATGTTCATGAGAATCCTTTGCTACGGTTAAAATGCCCTGTAAGGACCGAATGTACCCAAATAAGGCCAGCAACGCGGTATACACCGATCAACACCCGTACATGTCTCTCAACATAGTTACGGTTAAAATACCCTGCAAGGACCGAATGTACCCAAATAAGGCAAGCAACGGAGTATACACCGGAGTCGCCAGTGTCCGGACTGACTCGCCGGGATAAGCTGACAGACCACTTATTTCCAATTGCGTTGATGTCTAAGAAGGCGAGCTTCCCCTCTCCTAGGTGCTGGTGGTGGCTCCGAACAAAGGGCATGGCTTTGATCCGTATGGTAACTAAGTTGATCGTAGAATAGTCACGCGAGCCAGCAGTCAAACCGGCAGATGGTTAATGATAATGAGAATCCTTTGCTACGGTTAAAATTATCTGGCGCCTCTCGTTTCCCCAGCGACAAGTGTAGCCTCTCTCCTTCCTCACGCTGTTGGGTCCCAAAGGCCTACTACTATCCGACCACTGGGACAACCCATGGGTTATAAACGTGCTCGTTAAAGGTCAAGAGGAAATAACCATCTGGTCGTCGGGGATCCGATGGCATCGCCAGGTATTACGCCCCTCCATGAGAACAAAACAGCTCGGATAACGAAAAATGGTACGTGAAGCTCGGTATGTGGGCATTAGGCACATTGCTCTGTTTCTTGTAGTCTGGTATGAGGACAATGTGTGTCCCAACCGACAACTGGTTCAATGGCCACCCCGACCTAAACGACGTTTCCGCTACGATTGTGAGGACTTTCACAGGTTTTCGTCTGGCAGAATCATGTCCTTTCTACTTATCACCTCACCATGCATCTTACCCCGCGCGCAAACTTAGCAGTTCCGACCCTGAGGCCTACTCCTGCCGCAGGTCCAGACACATGGCCCCTGTGGATGCTGTGTACCTTGCTAACTGGCAAAGGTTAGCTCCCGTACCAATGCTAGAAGACGTGGCCAAGTCAAGTAATACCAACATTCGACAGAATCTATGATATAAGTGTGGTGTTTTGCGAATGGCATTGGCCGCCATACTTTTCCTACGGAAAGCTTAACTATCACTGTCTGTGTTATGAGCGCAAGGTGTGTCCCAATCGACAACTGGTTCAATGGCCACCCCGATCAAACGGGCTTCTCGACCTTTTTTTGTCCCAGGCTCAGGATTAGGCTAACAGATTAAGAGGGGTTTCAAGGAATAATTGACTTATGCGCGCGTCGACTATATTTTACTACCAGGCAGAATCATGTCCTTACTGCTTATCCAGGTCTTTTAAAGTTCGTTTCAGACCTTGGTCCCAAGCGACACCAAACGGAGAACTGTACGAATTGCTACCCAGTCCTACAAAAAAATGACTCAGTGGGCACAGGGGGTTCACAGTGTGATGAGTTAGCGTCCGCGAGAAAATATAATCTGTAAATCGGCCTCCGCCATTCAAGGGCCACCTATCAGCACCGAATTTTTGACTTAATACGTTCCTCACATTGTTAAGATGTTTAATCGCCATCGCATATACACAAAGGCCCCTGAAACTAATTAAGCAGACCTAAGAAAAAATGCCTCAGTGGGCACAGGGGGTTCACAGTGTCATCAGTTAGCGTCCGGGAGAAAATATAATCTGTGTAGCGAGGATTCTACATCAGACCGATAAATCTAATCGGCCCGGTATAAATTGTCGTACCGAACAACTTCCCCAAGAACGGTTCCCGTTATCACACTCATAGTATCTTACTAAACCACAGCAGGTGCAATATCGAAACACGTACTTTCCTGCCGCTAGTTTCGCTGGTGAGAGAATGTGATGCAATACTACTCAGGGGGCTAAGGAGCGGAGGTCGTTCTCTTTTAG
