+++ test/data/ref.fasta
@@ -1,0 +1,16 @@
+>1
+GATGACTCAGTCTTGTTGATTAGGCACCTCGGTATGTGGGCATTAGGCACATTGCTCTGTTTCTTGAAGT
+CTGGTATGAGCACAAGGTGTGTCCCAACCGACAACTGGTTCAATGGCCACCCCGACCTAAAGGACGTTTC
+CGCTTCGATTGTGAGGACTTTAACAGGTTTTCGTCTGGCAGAATCATGTCCTTACTGCTTATCCAGGTCT
+TTTAAAGTTCGTTTCAGACTTTGGTCCCAAGCGACTCCAAACGGAGAACCGTACGAATATCTACCCAGTC
+CTACGAAAAAATGACTCAGTGGGCACAGGGGGTTTACAGTGTGATCAGTTAGCGTCCGGGAGAAAATATA
+ATCTGTGTAGCGAGGATTCTATATCAGACCGATAAATCTAATCGGCCCGGTATAAATTCTCGTACCGAAC
+AACTTCTAACTTGGACGGTTGTACACTTACGGGAACTAGCAAACTCTAAGATAATAGAAGGCCTGAATCC
+GGCGTTACACCCATAGGGGAATAACGCCGAAATTGGTGGTTCTCGATAATTGCCAGTAATGCATCACAGC
+ATCTGCCTGGGTGGGGAATTGGGACAACCCTTGGGTTATAGACGTGCTCGTCAAAGGACAAGAGGAAATA
+CCCATCTGGTCATCGGGGATCCGATGGCATCGCCAGGTATTACGCCCCTCCATGAGAACAAAACAGCTCG
+GATAACGGTCAAACCGGCAGATGGTTAATGATCATGAGAATCCTTTGCTACGGTTAAAATACCCTGTAAG
+GACCGAATGTACCCAAATAAGGCAAGCAACGGAGTATACACCGGAGTCTCCAGTGTTCGGACTGACTCGC
+TGGGATAAGCTGACAGACGACTTATTGCCAATTGCGTTGATGTCTAAGAAGGCGAGCTTCCCCTCTCCTA
+GGTGCTGGTGGTGGCTCCGAACAAGGGGCTGACCTGCTCACCAGGTATTGTAGAGATCTGGCCATGGGTT
+TGACGGTTAATTCGAACAAATTTAGATGATTATTCCGTATTAGA
